move: CommentInfo 파일 따로 두기
