+++ DevFriends/DevFriends/Presentation/PostDetailScene/View/PostDetailCommentView.swift
@@ -8,15 +8,6 @@
 import UIKit
 import SnapKit
 
-struct CommentInfo: Hashable {
-    static func == (lhs: CommentInfo, rhs: CommentInfo) -> Bool {
-        return lhs.writerInfo == rhs.writerInfo && lhs.contents == rhs.contents
-    }
-    
-    let writerInfo: PostWriterInfo
-    let contents: String
-}
-
 final class PostDetailCommentView: UIView {
     private let stackView: UIStackView = {
         let stackView = UIStackView()
